move games main page files into own dir
